Change to SVG rendering for plotly PDF charts
